Improv(main): remove the useless usage of pending
Pending is an infinite loop, this means that if run() terminates the program will be stuck in an endless (unresponsive) loop: remove the usage of pending and replace it with a join on run so that if the program exits due to an error systemd can restart it; this also avoids using an unnecessary task.

Also avoid losing reference to the spawned Ctrl+C task.