Fix(Hardware Support): Passthrough Zotac Zone Touchpads
Users on SteamOS are reporting that touchpads do not work on the Zotac
Zone. Steamos-manager currently only creates a deck-uhid target so the
rel events produced have no target device to emit from. Don't grab the
touchpad anymore to reintroduce this functionality. This may need to be
changed in the future if touchpad firmware is updated to include an
abs event mode, though it would likely have a different report
descriptor and name at that point.