Fix(Hardware Support): Avoid using name as match for Ally/AllyX devices. - Combine Xbox and standard Ally device configs - Simplify evdev matches to not use name
Due to frequent changes to drivers upstream and OOT, there has been a lot
of volatility in the configuration for the Ally and Ally X devices. To
avoid further meddling by outside actors, drop any hid-asus, his-asus-ally,
or asus-ally-hid name attributes and bulk match all devices with the VID/PID
pairs.

While at it, reduce the number of device configs.